requirements.txt
py-cpuinfo>=7
psutil>=5
distro>=1
pyroute2>=0.5